pyproject.toml
[tool.poetry]
name = "social-app"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"
packages = [{include = "social_app"}]

[tool.poetry.dependencies]
python = "^3.9"
django = "^4.2.3"
djangorestframework = "^3.14.0"
pillow = "^10.0.0"
djangorestframework-simplejwt = "^5.2.2"
pep8 = "^1.7.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
